+++ packages/orq-rc/src/orq_ai_sdk/files.py
@@ -79,7 +79,7 @@ class Files(BaseSDK):
             hook_ctx=HookContext(
                 base_url=base_url or "",
                 operation_id="FileUpload",
-                oauth2_scopes=None,
+                oauth2_scopes=[],
                 security_source=get_security_from_env(
                     self.sdk_configuration.security, models.Security
                 ),
@@ -183,7 +183,7 @@ class Files(BaseSDK):
             hook_ctx=HookContext(
                 base_url=base_url or "",
                 operation_id="FileUpload",
-                oauth2_scopes=None,
+                oauth2_scopes=[],
                 security_source=get_security_from_env(
                     self.sdk_configuration.security, models.Security
                 ),
@@ -287,7 +287,7 @@ class Files(BaseSDK):
             hook_ctx=HookContext(
                 base_url=base_url or "",
                 operation_id="FileList",
-                oauth2_scopes=None,
+                oauth2_scopes=[],
                 security_source=get_security_from_env(
                     self.sdk_configuration.security, models.Security
                 ),
@@ -391,7 +391,7 @@ class Files(BaseSDK):
             hook_ctx=HookContext(
                 base_url=base_url or "",
                 operation_id="FileList",
-                oauth2_scopes=None,
+                oauth2_scopes=[],
                 security_source=get_security_from_env(
                     self.sdk_configuration.security, models.Security
                 ),
@@ -489,7 +489,7 @@ class Files(BaseSDK):
             hook_ctx=HookContext(
                 base_url=base_url or "",
                 operation_id="FileGet",
-                oauth2_scopes=None,
+                oauth2_scopes=[],
                 security_source=get_security_from_env(
                     self.sdk_configuration.security, models.Security
                 ),
@@ -587,7 +587,7 @@ class Files(BaseSDK):
             hook_ctx=HookContext(
                 base_url=base_url or "",
                 operation_id="FileGet",
-                oauth2_scopes=None,
+                oauth2_scopes=[],
                 security_source=get_security_from_env(
                     self.sdk_configuration.security, models.Security
                 ),
@@ -683,7 +683,7 @@ class Files(BaseSDK):
             hook_ctx=HookContext(
                 base_url=base_url or "",
                 operation_id="FileDelete",
-                oauth2_scopes=None,
+                oauth2_scopes=[],
                 security_source=get_security_from_env(
                     self.sdk_configuration.security, models.Security
                 ),
@@ -777,7 +777,7 @@ class Files(BaseSDK):
             hook_ctx=HookContext(
                 base_url=base_url or "",
                 operation_id="FileDelete",
-                oauth2_scopes=None,
+                oauth2_scopes=[],
                 security_source=get_security_from_env(
                     self.sdk_configuration.security, models.Security
                 ),