+++ packages/orq-rc/src/orq_ai_sdk/models/createchunkop.py
@@ -257,7 +257,7 @@ class CreateChunkResponseBody(BaseModel):
     updated_by_id: str
     r"""The unique identifier of the user who updated the chunk"""
 
-    id: Optional[str] = "chunk_01JQCTRB5641VM9Z2857QER9BA"
+    id: Optional[str] = "chunk_01JQHH0YGJR4J0WCEYH8TSWS56"
     r"""Unique identifier for the element"""
 
     enabled: Optional[bool] = True