+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -226,7 +226,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JQCTRB55NRQE9TA5CZ38G32M"
+        "01JQHH0YGGVQWD05R0W93SM8RA"
     )
     r"""The id of the resource"""