+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -86,7 +86,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JQCTRB547V5ST5RD1TFN2CWW"
+        "01JQHH0YGGFZQSZ9XSQPVH3GCY"
     )
     r"""The id of the resource"""