+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatapointop.py
@@ -384,5 +384,5 @@ class UpdateDatapointResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = dateutil.parser.isoparse("2025-03-27T22:14:48.594Z")
+    updated: Optional[datetime] = dateutil.parser.isoparse("2025-03-29T18:00:56.880Z")
     r"""The date and time the resource was last updated"""