+++ packages/orq-rc/src/orq_ai_sdk/prompts.py
@@ -79,7 +79,7 @@ class Prompts(BaseSDK):
             hook_ctx=HookContext(
                 base_url=base_url or "",
                 operation_id="GetAllPrompts",
-                oauth2_scopes=None,
+                oauth2_scopes=[],
                 security_source=get_security_from_env(
                     self.sdk_configuration.security, models.Security
                 ),
@@ -183,7 +183,7 @@ class Prompts(BaseSDK):
             hook_ctx=HookContext(
                 base_url=base_url or "",
                 operation_id="GetAllPrompts",
-                oauth2_scopes=None,
+                oauth2_scopes=[],
                 security_source=get_security_from_env(
                     self.sdk_configuration.security, models.Security
                 ),
@@ -286,7 +286,7 @@ class Prompts(BaseSDK):
             hook_ctx=HookContext(
                 base_url=base_url or "",
                 operation_id="CreatePrompt",
-                oauth2_scopes=None,
+                oauth2_scopes=[],
                 security_source=get_security_from_env(
                     self.sdk_configuration.security, models.Security
                 ),
@@ -389,7 +389,7 @@ class Prompts(BaseSDK):
             hook_ctx=HookContext(
                 base_url=base_url or "",
                 operation_id="CreatePrompt",
-                oauth2_scopes=None,
+                oauth2_scopes=[],
                 security_source=get_security_from_env(
                     self.sdk_configuration.security, models.Security
                 ),
@@ -487,7 +487,7 @@ class Prompts(BaseSDK):
             hook_ctx=HookContext(
                 base_url=base_url or "",
                 operation_id="GetOnePrompt",
-                oauth2_scopes=None,
+                oauth2_scopes=[],
                 security_source=get_security_from_env(
                     self.sdk_configuration.security, models.Security
                 ),
@@ -585,7 +585,7 @@ class Prompts(BaseSDK):
             hook_ctx=HookContext(
                 base_url=base_url or "",
                 operation_id="GetOnePrompt",
-                oauth2_scopes=None,
+                oauth2_scopes=[],
                 security_source=get_security_from_env(
                     self.sdk_configuration.security, models.Security
                 ),
@@ -731,7 +731,7 @@ class Prompts(BaseSDK):
             hook_ctx=HookContext(
                 base_url=base_url or "",
                 operation_id="UpdatePrompt",
-                oauth2_scopes=None,
+                oauth2_scopes=[],
                 security_source=get_security_from_env(
                     self.sdk_configuration.security, models.Security
                 ),
@@ -883,7 +883,7 @@ class Prompts(BaseSDK):
             hook_ctx=HookContext(
                 base_url=base_url or "",
                 operation_id="UpdatePrompt",
-                oauth2_scopes=None,
+                oauth2_scopes=[],
                 security_source=get_security_from_env(
                     self.sdk_configuration.security, models.Security
                 ),
@@ -985,7 +985,7 @@ class Prompts(BaseSDK):
             hook_ctx=HookContext(
                 base_url=base_url or "",
                 operation_id="DeletePrompt",
-                oauth2_scopes=None,
+                oauth2_scopes=[],
                 security_source=get_security_from_env(
                     self.sdk_configuration.security, models.Security
                 ),
@@ -1079,7 +1079,7 @@ class Prompts(BaseSDK):
             hook_ctx=HookContext(
                 base_url=base_url or "",
                 operation_id="DeletePrompt",
-                oauth2_scopes=None,
+                oauth2_scopes=[],
                 security_source=get_security_from_env(
                     self.sdk_configuration.security, models.Security
                 ),
@@ -1184,7 +1184,7 @@ class Prompts(BaseSDK):
             hook_ctx=HookContext(
                 base_url=base_url or "",
                 operation_id="ListPromptVersions",
-                oauth2_scopes=None,
+                oauth2_scopes=[],
                 security_source=get_security_from_env(
                     self.sdk_configuration.security, models.Security
                 ),
@@ -1291,7 +1291,7 @@ class Prompts(BaseSDK):
             hook_ctx=HookContext(
                 base_url=base_url or "",
                 operation_id="ListPromptVersions",
-                oauth2_scopes=None,
+                oauth2_scopes=[],
                 security_source=get_security_from_env(
                     self.sdk_configuration.security, models.Security
                 ),
@@ -1392,7 +1392,7 @@ class Prompts(BaseSDK):
             hook_ctx=HookContext(
                 base_url=base_url or "",
                 operation_id="GetPromptVersion",
-                oauth2_scopes=None,
+                oauth2_scopes=[],
                 security_source=get_security_from_env(
                     self.sdk_configuration.security, models.Security
                 ),
@@ -1499,7 +1499,7 @@ class Prompts(BaseSDK):
             hook_ctx=HookContext(
                 base_url=base_url or "",
                 operation_id="GetPromptVersion",
-                oauth2_scopes=None,
+                oauth2_scopes=[],
                 security_source=get_security_from_env(
                     self.sdk_configuration.security, models.Security
                 ),