+++ packages/orq-rc/RELEASES.md
@@ -949,3 +949,13 @@ Based on:
 - [python v3.3.0-rc.13] packages/orq-rc
 ### Releases
 - [PyPI v3.3.0-rc.13] https://pypi.org/project/orq-ai-sdk/3.3.0-rc.13 - packages/orq-rc
+
+## 2025-03-29 18:02:20
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.526.0 (2.563.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.3.0-rc.14] packages/orq-rc
+### Releases
+- [PyPI v3.3.0-rc.14] https://pypi.org/project/orq-ai-sdk/3.3.0-rc.14 - packages/orq-rc